refactor: extract api key file lock
